+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This change
 - Support for JSON format output.
 - Support for detecting foreign key relationships and organizing generated SQL accordingly.
 - Added option to use driver JARs on the classpath rather than requiring all SQL drivers to be compiled into the JAR.
+- Command line options to filter tables scanned by schema.
 
 ### Fixed
 - Passing `-h` or `--help` did not cause the usage information to be printed.